Add support for float values
Allow for floating point values, for more specific border sizes, including those between 0-1%

+++ python/src/argParser.py
@@ -12,7 +12,7 @@ def createConfig(argv):
             elif opt in ['-F']:
                 cParams["dir"] = arg
             elif opt in ['-b']:
-                cParams["border"] = int(arg)
+                cParams["border"] = float(arg)
             elif opt in ['-c']:
                 cParams["colour"] = arg
             elif opt in ['-l']:
